+++ Sources/CodexBar/StatusItemController+MenuWidthCache.swift
@@ -54,6 +54,7 @@ extension StatusItemController {
             canCheckForUpdates: self.updater.isAvailable,
             includeContextualActions: includeContextualActions,
             codexWorkspacesMenuEnabled: codexWorkspacesMenuEnabled,
+            isKeepingAwake: self.agentSessions.isKeepingAwake,
             agentSessionsEnabled: self.settings.agentSessionsEnabled,
             agentSessionLabelStyle: self.settings.agentSessionLabelStyle,
             agentSessionsHideUnreachableHosts: self.settings.agentSessionsHideUnreachableHosts,