+++ Sources/CodexBar/StatusItemController+Shutdown.swift
@@ -23,6 +23,8 @@ extension StatusItemController {
 
     private func cancelShutdownTasks() {
         self.agentSessions.stop()
+        self.store.credentialNotificationsStopped = true
+        self.store.retireCredentialNotifications()
         self.menuAppearanceObserver?.stop()
         self.menuAppearanceObserver = nil
         self.blinkTask?.cancel()